moves some methods to MButils
