+++ org.opentosca.bus/org.opentosca.bus.management.deployment.plugin.remote/src/main/java/org/opentosca/bus/management/deployment/plugin/remote/ManagementBusDeploymentPluginRemote.java
@@ -11,7 +11,7 @@ import org.apache.camel.Exchange;
 import org.apache.camel.Message;
 import org.opentosca.bus.management.deployment.plugin.IManagementBusDeploymentPluginService;
 import org.opentosca.bus.management.header.MBHeader;
-import org.opentosca.bus.management.service.impl.Constants;
+import org.opentosca.bus.management.Constants;
 import org.opentosca.bus.management.service.impl.collaboration.RequestSender;
 import org.opentosca.bus.management.service.impl.collaboration.model.CollaborationMessage;
 import org.opentosca.bus.management.service.impl.collaboration.model.KeyValueMap;
