+++ org.opentosca.bus/org.opentosca.bus.management.invocation.plugin.remote/src/main/java/org/opentosca/bus/management/invocation/plugin/remote/ManagementBusInvocationPluginRemote.java
@@ -13,7 +13,7 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import org.apache.camel.Exchange;
 import org.apache.camel.Message;
 import org.opentosca.bus.management.invocation.plugin.IManagementBusInvocationPluginService;
-import org.opentosca.bus.management.service.impl.Constants;
+import org.opentosca.bus.management.Constants;
 import org.opentosca.bus.management.service.impl.collaboration.RequestSender;
 import org.opentosca.bus.management.service.impl.collaboration.model.BodyType;
 import org.opentosca.bus.management.service.impl.collaboration.model.CollaborationMessage;
