+++ org.opentosca.bus/org.opentosca.bus.management/src/main/java/org/opentosca/bus/management/Constants.java
@@ -1,4 +1,4 @@
-package org.opentosca.bus.management.service.impl;
+package org.opentosca.bus.management;
 
 import javax.xml.namespace.QName;
